servers/python/RAG-memory-mcp/requirements.txt
fastmcp
chromadb
sentence-transformers
pydantic